+++ src/dgisim/character/enums.py
@@ -11,10 +11,10 @@ __all__ = [
 
 
 class CharacterSkill(Enum):
-    SKILL1 = 1  # typically the normal attack
-    SKILL2 = 2  # typically the 1st elemental skill
-    SKILL3 = 3  # typically the 2nd elemental skill
-    ELEMENTAL_BURST = 4
+    SKILL1 = 0  # typically the normal attack
+    SKILL2 = 1  # typically the 1st elemental skill
+    SKILL3 = 2  # typically the 2nd elemental skill
+    ELEMENTAL_BURST = 3
 
     def to_event_type(self) -> EventType:
         if self is CharacterSkill.SKILL1:
@@ -28,9 +28,9 @@ class CharacterSkill(Enum):
         raise NotImplementedError
 
 class CharacterSkillType(Enum):
-    NORMAL_ATTACK = 1
-    ELEMENTAL_SKILL = 2
-    ELEMENTAL_BURST = 3
+    NORMAL_ATTACK = 0
+    ELEMENTAL_SKILL = 1
+    ELEMENTAL_BURST = 2
 
     def is_normal_attack(self):
         return self is CharacterSkillType.NORMAL_ATTACK
@@ -42,22 +42,22 @@ class CharacterSkillType(Enum):
         return self is CharacterSkillType.ELEMENTAL_BURST
 
 class WeaponType(Enum):
-    BOW = 1
-    CATALYST = 2
-    CLAYMORE = 3
-    POLEARM = 4
-    SWORD = 5
-    NONE = 6
+    BOW = 0
+    CATALYST = 1
+    CLAYMORE = 2
+    POLEARM = 3
+    SWORD = 4
+    NONE = 5
 
 class Faction(Enum):
-    MONDSTADT = 1
-    LIYUE = 2
-    INAZUMA = 3
-    SUMERU = 4
-    FONTAINE = 5
-    NATLAN = 6
-    SNEZHNAYA = 7
-    FATUI = 8
-    MONSTER = 9
-    HILICHURL = 10
-    THE_EREMITES = 11
+    MONDSTADT = 0
+    LIYUE = 1
+    INAZUMA = 2
+    SUMERU = 3
+    FONTAINE = 4
+    NATLAN = 5
+    SNEZHNAYA = 6
+    FATUI = 7
+    MONSTER = 8
+    HILICHURL = 9
+    THE_EREMITES = 10