+++ src/dgisim/status/status.py
@@ -656,7 +656,7 @@ class Status:
             [self.__getattribute__(field.name)]
             for field in fields(self)
         ]))
-        ret_val = [encoding_plan.code_for(self)]
+        ret_val = [encoding_plan.encode_item(self)]
         for value in values:
             if isinstance(value, bool):
                 ret_val.append(1 if value else 0)
@@ -664,7 +664,7 @@ class Status:
                 ret_val.append(value)
             elif isinstance(value, Enum):
                 assert isinstance(value.value, int), value
-                ret_val.append(value.value)
+                ret_val.append(encoding_plan.encode_item(value))
             elif value is None:
                 ret_val.append(0)
             else: